+++ core/mount/mount_idmapped_utils_linux.go
@@ -1,5 +1,3 @@
-//go:build go1.23 && linux
-
 /*
    Copyright The containerd Authors.
 
@@ -29,17 +27,8 @@ import (
 )
 
 // getUsernsFD returns pinnable user namespace's file descriptor.
-//
-// NOTE: The GO runtime uses pidfd to handle subprocess since go1.23. However,
-// it has double close issue tracked by [1]. We can't use pidfd directly and
-// the GO runtime doesn't export interface to show if it's using pidfd or not.
-// So, we call `sys.SupportsPidFD` first and then use `os.Process` directly.
-//
-// [1]: https://github.com/golang/go/issues/68984
 func getUsernsFD(uidMaps, gidMaps []syscall.SysProcIDMap) (_ *os.File, retErr error) {
-	if !sys.SupportsPidFD() {
-		return nil, fmt.Errorf("failed to prevent pid reused issue because pidfd isn't supported")
-	}
+	var pidfd int
 
 	proc, err := os.StartProcess("/proc/self/exe", []string{"containerd[getUsernsFD]"}, &os.ProcAttr{
 		Sys: &syscall.SysProcAttr{
@@ -50,15 +39,26 @@ func getUsernsFD(uidMaps, gidMaps []syscall.SysProcIDMap) (_ *os.File, retErr er
 			// be in PTRACE_TRACEME mode before performing execve.
 			Ptrace:    true,
 			Pdeathsig: syscall.SIGKILL,
+			PidFD:     &pidfd,
 		},
 	})
 	if err != nil {
 		return nil, fmt.Errorf("failed to start noop process for unshare: %w", err)
 	}
 
-	defer func() {
+	if pidfd == -1 || !sys.SupportsPidFD() {
 		proc.Kill()
 		proc.Wait()
+		return nil, fmt.Errorf("failed to prevent pid reused issue because pidfd isn't supported")
+	}
+
+	pidFD := os.NewFile(uintptr(pidfd), "pidfd")
+	defer func() {
+		unix.PidfdSendSignal(int(pidFD.Fd()), unix.SIGKILL, nil, 0)
+
+		pidfdWaitid(pidFD)
+
+		pidFD.Close()
 	}()
 
 	// NOTE:
@@ -78,8 +78,14 @@ func getUsernsFD(uidMaps, gidMaps []syscall.SysProcIDMap) (_ *os.File, retErr er
 	// Ensure the child process is still alive. If the err is ESRCH, we
 	// should return error because we can't guarantee the usernsFD and
 	// u[g]idmapFile are valid. It's safe to return error and retry.
-	if err := proc.Signal(syscall.Signal(0)); err != nil {
+	if err := unix.PidfdSendSignal(int(pidFD.Fd()), 0, nil, 0); err != nil {
 		return nil, fmt.Errorf("failed to ensure child process is alive: %w", err)
 	}
 	return usernsFD, nil
+}
+
+func pidfdWaitid(pidFD *os.File) error {
+	return sys.IgnoringEINTR(func() error {
+		return unix.Waitid(unix.P_PIDFD, int(pidFD.Fd()), nil, unix.WEXITED, nil)
+	})
 }
